src/test/java/com/saucelabs/saucerest/integration/BuildsEndpointTest.java
package com.saucelabs.saucerest.integration;

import com.saucelabs.saucerest.DataCenter;
import com.saucelabs.saucerest.JobSource;
import com.saucelabs.saucerest.SauceREST;
import com.saucelabs.saucerest.api.BuildsEndpoint;
import com.saucelabs.saucerest.model.builds.Build;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.EnumSource;

import java.io.IOException;
import java.util.List;

import static org.junit.jupiter.api.Assertions.assertTrue;

public class BuildsEndpointTest {
    @ParameterizedTest
    @EnumSource(value = DataCenter.class, names = {"US_EAST"}, mode = EnumSource.Mode.EXCLUDE)
    public void lookupBuildsTest(DataCenter dataCenter) throws IOException {
        SauceREST sauceREST = new SauceREST(dataCenter);
        BuildsEndpoint buildsEndpoint = sauceREST.getBuildsEndpoint();

        List<Build> builds = buildsEndpoint.lookupBuilds(JobSource.VDC);

        assertTrue(builds.size() > 0);
    }
}